Rough Draft of a Refactor with using Fragments Will clean up in the next commit.
MapViewFragment.kt
-Fragment class for hosting the map view and all its functions

MainActivity.kt, activity_main.xml
-Commented out the old implementation and replaced with the navigation and fragment approach

map.xml
-Navigation view for the mapView fragment

app_navigation.xml
-Navigation view for the entire application
