+++ app/src/main/res/layout/activity_main.xml
@@ -1,64 +1,80 @@
 <?xml version="1.0" encoding="utf-8"?>
-<layout
-    xmlns:android="http://schemas.android.com/apk/res/android"
+<LinearLayout
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools">
+    xmlns:tools="http://schemas.android.com/tools"
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    tools:context=".MainActivity">
 
-    <androidx.drawerlayout.widget.DrawerLayout
+    <androidx.fragment.app.FragmentContainerView
+        android:id="@+id/nav_host_container"
+        android:name="androidx.navigation.fragment.NavHostFragment"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:id="@+id/drawer_layout">
+        app:defaultNavHost="true"
+        app:navGraph="@navigation/app_navigation"/>
 
-        <LinearLayout
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:orientation="vertical">
-            <include
-                layout="@layout/toolbar"
-                android:layout_height="wrap_content"
-                android:layout_width="match_parent"/>
-            <FrameLayout
-                android:id="@+id/frameLayout"
-                android:layout_width="match_parent"
-                android:layout_height="match_parent">
-                <androidx.fragment.app.FragmentContainerView
-                    android:layout_width="match_parent"
-                    android:layout_height="match_parent"
-                    android:id="@+id/nav_host_fragment"
-
-                    />
-
-            </FrameLayout>
-        </LinearLayout>
+</LinearLayout>
 
 
+<!--<layout-->
+<!--    xmlns:android="http://schemas.android.com/apk/res/android"-->
+<!--    xmlns:app="http://schemas.android.com/apk/res-auto"-->
+<!--    xmlns:tools="http://schemas.android.com/tools">-->
 
-        <com.google.android.material.navigation.NavigationView
-            android:id="@+id/navigationView"
-            android:layout_width="wrap_content"
-            android:layout_height="match_parent"
-            android:layout_gravity="start"
-            android:background="@android:color/background_light"
-            app:headerLayout="@layout/nav_header"
-            app:menu="@menu/drawer_view"/>
+<!--    <androidx.drawerlayout.widget.DrawerLayout-->
+<!--        android:layout_width="match_parent"-->
+<!--        android:layout_height="match_parent"-->
+<!--        android:id="@+id/drawer_layout">-->
 
-        <com.arcgismaps.mapping.view.MapView
-            android:id="@+id/mapView"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent" />
-
-<!--        <androidx.constraintlayout.widget.ConstraintLayout-->
-<!--            xmlns:tools="http://schemas.android.com/tools"-->
+<!--        <LinearLayout-->
 <!--            android:layout_width="match_parent"-->
 <!--            android:layout_height="match_parent"-->
-<!--            tools:context=".MainActivity">-->
-<!--            <com.arcgismaps.mapping.view.MapView-->
-<!--                android:id="@+id/mapView" -->
-<!--                android:layout_width="match_parent" -->
-<!--                android:layout_height="match_parent" />-->
-<!--        </androidx.constraintlayout.widget.ConstraintLayout>-->
-    </androidx.drawerlayout.widget.DrawerLayout>
-</layout>
+<!--            android:orientation="vertical">-->
+<!--            <include-->
+<!--                layout="@layout/toolbar"-->
+<!--                android:layout_height="wrap_content"-->
+<!--                android:layout_width="match_parent"/>-->
+<!--            <FrameLayout-->
+<!--                android:id="@+id/frameLayout"-->
+<!--                android:layout_width="match_parent"-->
+<!--                android:layout_height="match_parent">-->
+<!--                <androidx.fragment.app.FragmentContainerView-->
+<!--                    android:layout_width="match_parent"-->
+<!--                    android:layout_height="match_parent"-->
+<!--                    android:id="@+id/nav_host_fragment"/>-->
+<!--            </FrameLayout>-->
+<!--        </LinearLayout>-->
+
+
+
+<!--        <com.google.android.material.navigation.NavigationView-->
+<!--            android:id="@+id/navigationView"-->
+<!--            android:layout_width="wrap_content"-->
+<!--            android:layout_height="match_parent"-->
+<!--            android:layout_gravity="start"-->
+<!--            android:background="@android:color/background_light"-->
+<!--            app:headerLayout="@layout/nav_header"-->
+<!--            app:menu="@menu/drawer_view"/>-->
+
+<!--        <com.arcgismaps.mapping.view.MapView-->
+<!--            android:id="@+id/mapView"-->
+<!--            android:layout_width="match_parent"-->
+<!--            android:layout_height="match_parent" />-->
+
+<!--&lt;!&ndash;        <androidx.constraintlayout.widget.ConstraintLayout&ndash;&gt;-->
+<!--&lt;!&ndash;            xmlns:tools="http://schemas.android.com/tools"&ndash;&gt;-->
+<!--&lt;!&ndash;            android:layout_width="match_parent"&ndash;&gt;-->
+<!--&lt;!&ndash;            android:layout_height="match_parent"&ndash;&gt;-->
+<!--&lt;!&ndash;            tools:context=".MainActivity">&ndash;&gt;-->
+<!--&lt;!&ndash;            <com.arcgismaps.mapping.view.MapView&ndash;&gt;-->
+<!--&lt;!&ndash;                android:id="@+id/mapView" &ndash;&gt;-->
+<!--&lt;!&ndash;                android:layout_width="match_parent" &ndash;&gt;-->
+<!--&lt;!&ndash;                android:layout_height="match_parent" />&ndash;&gt;-->
+<!--&lt;!&ndash;        </androidx.constraintlayout.widget.ConstraintLayout>&ndash;&gt;-->
+<!--    </androidx.drawerlayout.widget.DrawerLayout>-->
+<!--</layout>-->
 
 
     <!-- <com.example.cs567_3d_ui_project.qgis_map.QGisMapViewContainer
