+++ app/src/main/res/navigation/map.xml
@@ -1,0 +1,15 @@
+<?xml version="1.0" encoding="utf-8"?>
+<navigation
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/map"
+    app:startDestination="@+id/mapViewFragment">
+
+    <fragment
+        android:id="@+id/mapViewFragment"
+        android:name="com.example.cs567_3d_ui_project.fragments.MapViewFragment"
+        tools:layout="@layout/fragment_map_view">
+    </fragment>
+
+</navigation>
